# book-network/src/main/java/com/mukund/booknetwork/book/FeedbackResponse.java
package com.mukund.booknetwork.book;


import lombok.*;

@Getter
@Setter
@AllArgsConstructor
@NoArgsConstructor
@Builder
public class FeedbackResponse {
    private Double note;
    private String comment;
    private boolean ownFeedback;


}